Revert Back to H-SV
The reason I use `float` here is to make it possible to create an RTL
version of the color picker by simply replaces the `float` direction
from `right` to `left` and vice-versa. Although, I am not in a urge to
add such feature as this is really easy to achive using CSS:

~~~ .css
[dir="rtl"] .color-picker-h {float:left}
[dir="rtl"] .color-picker-sv {float:right}
~~~

I also revert back the markup order so that the H slider comes first
before SV so that it will correctly says H-SV instead of SV-H.

I hope you don’t mind.