Drag color and release mouse outside closes picker
If I press and drag color within bounds of color picker it stays open.

if I press and drag color and then release mouse outside of bounds of picker it closes, which is counter intuitive.

Color picker should track when I pressed initially, if it was inside picker, then releasing mouse outside of bounds should not cause picker to close. If I initially pressed outside of bounds and then released outside, it closes picker.